refactor: update axios interceptors and error handling logic
