+++ src/config/axios/constants/index.ts
@@ -12,3 +12,13 @@ export const METHODS = {
 } as const;
 
 export type MethodType = (typeof METHODS)[keyof typeof METHODS];
+
+export const RETRY_CONFIG = {
+  MAX_RETRIES: 3,
+  BASE_DELAY_MS: 1000,
+} as const;
+
+export const JWT_REFRESH_CONFIG = {
+  LOGIN_REDIRECT_PATH: '/auth/login',
+  REFRESH_TOKEN_ENDPOINT: '/auth/refresh',
+} as const;
